# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LazyCat Community App Store API
  version: 0.1.0
  description: REST API for the self-hosted LazyCat LPK app store server.
servers:
  - url: http://localhost:8080
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/client/v1/sources:
    get:
      tags: [Client]
      summary: List software sources stored by the standalone client
      description: Client endpoints use the optional x-hc-user-id header supplied by the LazyCat runtime. When it is absent, data is scoped to the local user.
      security: []
      responses:
        "200":
          description: Source subscriptions.
          content:
            application/json:
              schema:
                type: object
                required: [sources]
                properties:
                  sources:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClientSource"
    post:
      tags: [Client]
      summary: Add a software source to the standalone client
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientSourceInput"
      responses:
        "201":
          description: Source created.
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source:
                    $ref: "#/components/schemas/ClientSource"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/client/v1/sources/{id}:
    patch:
      tags: [Client]
      summary: Update a standalone client source
      security: []
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientSourceInput"
      responses:
        "200":
          description: Source updated.
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source:
                    $ref: "#/components/schemas/ClientSource"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Client]
      summary: Delete a standalone client source and its cached apps
      security: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Source deleted.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "404":
          $ref: "#/components/responses/Error"
  /api/client/v1/sources/{id}/sync:
    post:
      tags: [Client]
      summary: Sync one standalone client source
      security: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Source sync result.
          content:
            application/json:
              schema:
                type: object
                required: [source]
                properties:
                  source:
                    $ref: "#/components/schemas/ClientSource"
        "401":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
...
  /api/client/v1/apps:
    get:
      tags: [Client]
      summary: List cached apps from synced client sources
      security: []
      parameters:
        - name: sourceId
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cached source apps.
          content:
            application/json:
              schema:
                type: object
                required: [apps]
                properties:
                  apps:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClientSourceApp"
  /api/client/v1/apps/{id}:
    get:
      tags: [Client]
      summary: Get one cached client source app
      security: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Cached source app.
          content:
            application/json:
              schema:
                type: object
                required: [app]
                properties:
                  app:
                    $ref: "#/components/schemas/ClientSourceApp"
        "404":
          $ref: "#/components/responses/Error"
  /api/client/v1/apps/{id}/versions:
    get:
      tags: [Client]
      summary: List versions for one cached client source app
      security: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Versions available from the synced source feed.
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClientVersion"
        "404":
          $ref: "#/components/responses/Error"
  /api/client/v1/installed:
    get:
      tags: [Client]
      summary: Query apps installed on the current LazyCat device
      description: This endpoint is local to the standalone client. The server storefront never receives or stores this device inventory.
      security: []
      responses:
        "200":
          description: Installed LazyCat applications.
          content:
            application/json:
              schema:
                type: object
                required: [apps]
                properties:
                  apps:
                    type: array
                    items:
                      $ref: "#/components/schemas/InstalledApplication"
        "502":
          $ref: "#/components/responses/Error"
  /api/client/v1/install:
    post:
      tags: [Client]
      summary: Install a cached source app through the LazyCat Go SDK
      description: If version is omitted, the latest source version is installed. Supplying version lets the client install an older synced version for rollback.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInstallRequest"
      responses:
        "200":
          description: Install task result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInstallResult"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/client/v1/history:
    get:
      tags: [Client]
      summary: List install attempts recorded by the standalone client
      security: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Install history entries.
          content:
            application/json:
              schema:
                type: object
                required: [history]
                properties:
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/InstallHistoryEntry"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lazycat_store_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              nullable: true
...
    ClientSource:
      type: object
      required: [id, name, url, password, mirror, lastAppCount, lastInstallableCount]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
          format: uri
        password:
          type: string
          description: Source feed password stored locally by the standalone client.
        mirror:
          type: string
          description: Optional mirror prefix used by the client for GitHub download URLs.
        lastSync:
          type: string
          format: date-time
        lastError:
          type: string
        lastErrorCode:
          type: string
        lastAppCount:
          type: integer
        lastInstallableCount:
          type: integer
    ClientSourceInput:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        password:
          type: string
        mirror:
          type: string
    ClientVersion:
      type: object
      required: [version, downloadUrl, sha256, size]
      properties:
        version:
          type: string
        downloadUrl:
          type: string
          format: uri
        upstreamDownloadUrl:
          type: string
          format: uri
        sourceType:
          type: string
        sha256:
          type: string
        size:
          type: integer
          format: int64
    ClientSourceApp:
      type: object
      required: [id, sourceId, sourceName, packageId, name, slug, summary, installProtected, versions]
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        sourceName:
          type: string
        packageId:
          type: string
          description: LazyCat package identity used to match installed apps and install history.
        name:
          type: string
        slug:
          type: string
        summary:
          type: string
        category:
          type: string
        installProtected:
          type: boolean
        latestVersion:
          $ref: "#/components/schemas/ClientVersion"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/ClientVersion"
...
    InstalledApplication:
      type: object
      required: [appid]
      properties:
        appid:
          type: string
          description: Package ID reported by the LazyCat runtime.
        title:
          type: string
        version:
          type: string
        status:
          type: string
        instanceStatus:
          type: string
        icon:
          type: string
    ClientInstallRequest:
      type: object
      required: [appId]
      properties:
        appId:
          type: integer
          description: Cached client source app ID.
        version:
          type: string
          description: Optional target version. Omit to install the latest synced version.
        installPassword:
          type: string
        name:
          type: string
          deprecated: true
          description: Ignored by normal client installs; the server derives install metadata from the cached source app.
        pkgId:
          type: string
          deprecated: true
        downloadUrl:
          type: string
          deprecated: true
        sha256:
          type: string
          deprecated: true
    ClientInstallResult:
      type: object
      required: [mode]
      properties:
        mode:
          type: string
        taskId:
          type: string
        status:
          type: string
        detail:
          type: string
    InstallHistoryEntry:
      type: object
      required: [id, packageId, appName, result, createdAt]
      properties:
        id:
          type: integer
        sourceId:
          type: integer
          nullable: true
        sourceAppId:
          type: integer
          nullable: true
        sourceName:
          type: string
        packageId:
          type: string
        appName:
          type: string
        version:
          type: string
        result:
          type: string
          enum: [SUCCESS, FAILED]
        downloadUrl:
          type: string
        sha256:
          type: string
        error:
          type: string
        createdAt:
          type: string
          format: date-time
...